--- a/src/app/modules/angular-slickgrid/formatters/__tests__/dateTimeIsoFormatter.spec.ts
+++ b/src/app/modules/angular-slickgrid/formatters/__tests__/dateTimeIsoFormatter.spec.ts
@@ -16,7 +16,7 @@ describe('the DateTimeIso Formatter', () => {
 
   it('should provide a dateIso formatted input and return a formatted date value without time when valid date value is provided', () => {
     const value = '2019-05-01 02:36:07';
-    const result = Formatters.dateTimeIso(0, 0, value, { input: 'dateIso' } as unknown as Column, {}, {});
+    const result = Formatters.dateTimeIso(0, 0, value, { type: 'dateIso' } as unknown as Column, {}, {});
     expect(result).toBe('2019-05-01 02:36:07');
   });