--- a/src/app/modules/angular-slickgrid/sorters/dateUtilities.ts
+++ b/src/app/modules/angular-slickgrid/sorters/dateUtilities.ts
@@ -6,18 +6,18 @@ const moment = moment_; // patch to fix rollup "moment has no default export" is
 export function compareDates(value1: any, value2: any, sortDirection: number, sortColumn: Column, gridOptions: GridOption, format: string | moment_.MomentBuiltinFormat, strict?: boolean) {
   let diff = 0;
   const checkForUndefinedValues = sortColumn && sortColumn.valueCouldBeUndefined || gridOptions && gridOptions.cellValueCouldBeUndefined || false;
+  const date1 = moment(value1, format, strict);
+  const date2 = moment(value2, format, strict);
 
-  if (value1 === null || value1 === '' || (checkForUndefinedValues && value1 === undefined) || !moment(value1, format, strict).isValid()) {
+  if (value1 === null || value1 === '' || (checkForUndefinedValues && value1 === undefined) || !date1.isValid()) {
     diff = -1;
-  } else if (value2 === null || value2 === '' || (checkForUndefinedValues && value2 === undefined) || !moment(value2, format, strict).isValid()) {
+  } else if (value2 === null || value2 === '' || (checkForUndefinedValues && value2 === undefined) || !date2.isValid()) {
     diff = 1;
   } else {
-    const date1 = moment(value1, format, strict);
-    const date2 = moment(value2, format, strict);
-    diff = parseInt(date1.format('X'), 10) - parseInt(date2.format('X'), 10);
+    diff = date1.valueOf() < date2.valueOf() ? -1 : 1;
   }
 
-  return sortDirection * (diff === 0 ? 0 : (diff > 0 ? 1 : -1));
+  return sortDirection * diff;
 }
 
 /** From a FieldType, return the associated date Sorter */