Here's a quick perf date sorting test before & after code change with 50k rows and 8 columns
_note: the date are all in ISO format (yyyy-mm-dd), it might be even bigger differences with non-ISO format_

### BEFORE
#### 1st pass
sorting ASC: 2255.841064453125 ms
sorting DESC: 676.7099609375 ms
sorting ASC: 644.962646484375 ms
sorting DESC: 671.908935546875 ms

#### 2nd pass
sorting ASC: 1981.27978515625 ms
sorting DESC: 692.999755859375 ms
sorting ASC: 673.643798828125 ms
sorting DESC: 670.118896484375 ms

### AFTER

#### 1st pass
sorting ASC: 1059.22802734375 ms
sorting DESC: 126.10400390625 ms
sorting ASC: 336.77685546875 ms
sorting DESC: 143.007080078125 ms

#### 2nd pass
sorting ASC: 903.991943359375 ms
sorting DESC: 143.51708984375 ms
sorting ASC: 374.537109375 ms
sorting DESC: 117.73583984375 ms